Factor out server logic into separate files
